+++ api/media.js
@@ -7,7 +7,8 @@ const crypto = require('crypto');
 const jwt = require('jsonwebtoken');
 const koaRoute = require('koa-route');
 const parse = require('co-busboy');
-const fs = require("fs");
+const createMedia = require("../service/createMediaService");
+const saveFile = require("../service/saveFileService");
 const app = new Koa();
 
 //fileds
@@ -16,47 +17,13 @@ const LINKS = "links";
 const DESCRIPTION = "description";
 const FOLDER_ID = "FolderId";
 
-const PATH_TO_FILES = './files/';
-
-const text = ["text/cmd",
-"text/css",
-"text/csv",
-"text/html",
-"text/javascript",
-"text/plain",
-"text/php",
-"text/xml"];
-
-const video = ["video/mpeg",
-"video/mp4",
-"video/ogg",
-"video/quicktime",
-"video/webm",
-"video/x-ms-wmv",
-"video/x-flv",
-"video/3gpp",
-"video/3gpp2"];
-
-const image = ["image/gif",
-"image/jpeg",
-"image/pjpeg",
-"image/png",
-"image/svg+xml",
-"image/tiff",
-"image/vnd.microsoft.icon",
-"image/vnd.wap.wbmp"];
-
 app.use(koaRoute.post('/', async (ctx, next) => {
   if (!ctx.request.is('multipart/*')){
     next();
     return;
   }
 
-  var media = {
-     name : '',
-     links:'',
-     description:''
-  };
+  var media = { name : '', links:'', description:'' };
 
   var parts = parse(ctx);
   var part;
@@ -67,9 +34,9 @@ app.use(koaRoute.post('/', async (ctx, next) => {
       console.log(part[0]);
       if(FOLDER_ID == part[0]){
         folderId = part[1];
-        if(!isThisUserCanWorkWithFolder(folderId, ctx.state.user.id)){
-          //return 403
-          ctx.throw(403);
+        if(!isThisUserCanWorkWithFolder(folderId, // folder id
+            ctx.state.user.id)){
+          ctx.throw(403); //return 403 forbidden
           next();
           return;
         }
@@ -77,9 +44,9 @@ app.use(koaRoute.post('/', async (ctx, next) => {
         fillField(ctx, media, part);
       }
     } else {
-      var resukt = await writeFile(ctx, part, media);
+      var resukt = await saveFile.saveMediaFile(part, media);
       if(!resukt){
-        console.log("400 tyhpe error");
+        console.log("400 type error");
           ctx.throw(400, 'there\'s not completely data');
           next();
           return;
@@ -87,18 +54,18 @@ app.use(koaRoute.post('/', async (ctx, next) => {
     }
   }
 
-  if(!isStructureFull(media) || !folderId){
-    if(media.file){//remove file
-      fs.unlinkSync(media.file);
-    }
-    //return 400
-    ctx.throw(400, 'there\'s not completely data');
-  } else {
-    //all good save to db and return 201
-    var folder = await db.Folder.findOne({where:{id: folderId}});
-    await folder.addMedium(await db.Media.create(media));
+  try{
+    console.log("createMedia.saveMedia(media, folderId)");
+    await createMedia.saveMedia(media, folderId);
+      console.log("end ");
     ctx.status = 201;
     ctx.body = 'data was added';
+  }catch(err){
+    console.log("err.message");
+    console.log(err.message);
+    if(err.message == "data_isnt_full")
+      ctx.throw(400, 'there\'s not completely data');
+
   }
   next();
 }));
@@ -107,11 +74,7 @@ app.use(koaRoute.post('/', async (ctx, next) => {
 app.use(koaRoute.get('/:id', async (ctx, mediaId, next) => {
   let media = await db.Media.findOne({where:{id: mediaId}});
   ctx.status = 200;
-  if(media){
-    ctx.body = media;
-  } else {
-    ctx.body = "{}";
-  }
+  ctx.body = (media) ? media : "{}" ;
   next();
 }));
 
@@ -121,39 +84,6 @@ function getNextPart(parts){
   });
 }
 
-async function writeFile(ctx, part, media){
-  var d = new Date();
-  var n = d.getTime();
-  var pathToFile = PATH_TO_FILES + n;
-  media.file = pathToFile;
-  var stream = fs.createWriteStream(pathToFile);
-  var closeStreamPromise = new Promise(ok => {
-      part.on('end', function(){ok()})
-  });
-  part.pipe(stream);
-  await closeStreamPromise;
-  var closeMyStreamPromise = new Promise(ok => {
-      stream.close(function(){ok()})
-  });
-  await closeMyStreamPromise;
-  var type = getType(part.mime);
-  if(!type){
-    return false;
-  }
-  media.type = type;
-  return true;
-}
-
-function getType(mime){
-  if(video.indexOf(mime) > -1){
-    return "video";
-  } else if(image.indexOf(mime) > -1){
-    return "image";
-  } else if(text.indexOf(mime) > -1){
-    return "text";
-  }
-  return null;
-}
 function fillField(ctx, media, part){
   var value = part[1];
   switch (part[0]) {
@@ -170,9 +100,7 @@ function fillField(ctx, media, part){
     default:
   }
 }
-function isStructureFull(media){
-    return media.name && media.links && media.file;
-}
+
 async function isThisUserCanWorkWithFolder(folderId, userId) {
   var currentFolder = await db.Folder.findOne({where:{id: folderId}});
   var user = await db.User.findOne({where: { id: userId }} );
